vcs-svn: improve support for reading large files
Move from uint32_t to off_t as the fundamental unit of length used by
the line_buffer library.  Performance would get worse if anything but
I think it's worth it for support of deltas that need to skip large
pieces (> 4 GiB).

Exception: buffer_read_string still takes a uint32_t, since it keeps
its result in an in-core obj_pool.

Callers still have to be updated to take advantage of this.

+++ vcs-svn/line_buffer.c
@@ -66,7 +66,7 @@ char *buffer_read_string(uint32_t len)
 	return ferror(infile) ? NULL : s;
 }
 
-void buffer_copy_bytes(uint32_t len)
+void buffer_copy_bytes(off_t len)
 {
 	uint32_t in;
 	while (len > 0 && !feof(infile) && !ferror(infile)) {
@@ -81,7 +81,7 @@ void buffer_copy_bytes(uint32_t len)
 	}
 }
 
-void buffer_skip_bytes(uint32_t len)
+void buffer_skip_bytes(off_t len)
 {
 	uint32_t in;
 	while (len > 0 && !feof(infile) && !ferror(infile)) {
