+++ vcs-svn/line_buffer.h
@@ -5,8 +5,8 @@ int buffer_init(const char *filename);
 int buffer_deinit(void);
 char *buffer_read_line(void);
 char *buffer_read_string(uint32_t len);
-void buffer_copy_bytes(uint32_t len);
-void buffer_skip_bytes(uint32_t len);
+void buffer_copy_bytes(off_t len);
+void buffer_skip_bytes(off_t len);
 void buffer_reset(void);
 
 #endif
